Move Pair class to tool module.
