+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/InventoryService.java
@@ -26,7 +26,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.axelor.apps.Pair;
+import com.axelor.apps.tool.Pair;
 import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.IAdministration;
 import com.axelor.apps.base.db.Product;
