+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/MrpLineService.java
@@ -21,7 +21,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
-import com.axelor.apps.Pair;
+import com.axelor.apps.tool.Pair;
 import com.axelor.apps.base.db.Partner;
 import com.axelor.apps.base.db.Product;
 import com.axelor.apps.purchase.db.PurchaseOrder;
